requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2

# Experiment Tracking
mlflow==2.9.1
optuna==3.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlite3  # Built-in with Python

# Jupyter notebooks
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
httpx==0.25.2